Restructured preview XRI Integration & fixed build error
