+++ Packages/Tracking Preview/XR Interaction Toolkit Integration/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 109ec17e159371f498cd47cecab5faa0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
