+++ Packages/Tracking Preview/XR Interaction Toolkit Integration/Editor/Ultraleap.Tracking.XRInteractionToolkit.Editor.asmdef
@@ -1,0 +1,23 @@
+{
+    "name": "Leap.Unity.Preview.XRInteractionToolkit.Editor",
+    "rootNamespace": "",
+    "references": [
+        "GUID:2ee0542102ac1084e8d9a910885c1e81",
+        "GUID:fe685ec1767f73d42b749ea8045bfe43",
+        "GUID:75469ad4d38634e559750d17036d5f7c",
+        "GUID:db4e617c04f44c144bc19216a63dd19b",
+        "GUID:cec9848704c49724192ff47f9b3e9156",
+        "GUID:506905c0b578be040956cc0439c8ab8b"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
