+++ Packages/Tracking Preview/XR Interaction Toolkit Integration/Editor/Ultraleap.Tracking.XRInteractionToolkit.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5bc6d8c5edc6f124ebe32048e95d0de5
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
