+++ Packages/Tracking Preview/XR Interaction Toolkit Integration/Runtime.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61ae4e5fd106c5c4eb6db3f7ec8fe624
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
