+++ Packages/Tracking Preview/XR Interaction Toolkit Integration/Runtime/FarFieldInteractions/Scripts/RotationDeadzone.cs.meta
