+++ Packages/Tracking Preview/XR Interaction Toolkit Integration/Runtime/Tracked Hands Ultraleap.prefab.meta
