+++ Packages/Tracking Preview/XR Interaction Toolkit Integration/Runtime/TrackedHandsController.cs.meta
